Uno.Threading: reorganize files
