+++ lib/Uno.Threading/Uno.Threading.unoproj
@@ -18,13 +18,13 @@
     "SyncDispatcher.uno:Source",
     "ThreadPool.uno:Source",
     "Tasks/Tasks.uno:Source",
-    "Implementation/Posix/posix_mutex.cpp:CSource:UNIX && !PInvoke",
-    "Implementation/Posix/posix_mutex.h:CHeader:UNIX && !PInvoke",
-    "Implementation/Posix/posix_reset_event.cpp:CSource:UNIX && !PInvoke",
-    "Implementation/Posix/posix_reset_event.h:CHeader:UNIX && !PInvoke",
-    "Implementation/Posix/posix_semaphore.cpp:CSource:UNIX && !PInvoke",
-    "Implementation/Posix/posix_semaphore.h:CHeader:UNIX && !PInvoke",
-    "Implementation/Posix/PthreadHelpers.uno:Source",
-    "Implementation/Win32/Win32Helpers.uno:Source",
+    "uThread/posix_mutex.cpp:CSource:UNIX && !PInvoke",
+    "uThread/posix_mutex.h:CHeader:UNIX && !PInvoke",
+    "uThread/posix_reset_event.cpp:CSource:UNIX && !PInvoke",
+    "uThread/posix_reset_event.h:CHeader:UNIX && !PInvoke",
+    "uThread/posix_semaphore.cpp:CSource:UNIX && !PInvoke",
+    "uThread/posix_semaphore.h:CHeader:UNIX && !PInvoke",
+    "PthreadHelpers.uno:Source",
+    "Win32Helpers.uno:Source",
   ]
 }
