+++ lib/Uno.Threading/uThread/posix_mutex.cpp
@@ -1,4 +1,4 @@
-#include <Implementation/Posix/posix_mutex.h>
+#include <uThread/posix_mutex.h>
 #include <pthread.h>
 #include <errno.h>
 #include <uBase/Time.h>
